requirements.txt
asgiref==3.5.2 ; python_version >= "3.8" and python_version < "4.0"
backports-zoneinfo==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
black==22.10.0 ; python_version >= "3.8" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4"
charset-normalizer==2.1.1 ; python_version >= "3.8" and python_version < "4"
click==8.1.3 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.5 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
dj-database-url==1.2.0 ; python_version >= "3.8" and python_version < "4.0"
django-cors-headers==3.13.0 ; python_version >= "3.8" and python_version < "4.0"
django-environ==0.9.0 ; python_version >= "3.8" and python_version < "4"
django==4.0.8 ; python_version >= "3.8" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.8" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4"
mypy-extensions==0.4.3 ; python_version >= "3.8" and python_version < "4.0"
pathspec==0.10.1 ; python_version >= "3.8" and python_version < "4.0"
platformdirs==2.5.2 ; python_version >= "3.8" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.8" and python_version < "4.0"
pytz==2022.6 ; python_version >= "3.8" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.8" and python_version < "4"
sentry-sdk==1.14.0 ; python_version >= "3.8" and python_version < "4.0"
setuptools==66.1.1 ; python_version >= "3.8" and python_version < "4.0"
sqlparse==0.4.3 ; python_version >= "3.8" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.8" and python_full_version < "3.11.0a7"
typing-extensions==4.4.0 ; python_version >= "3.8" and python_version < "3.10"
tzdata==2022.2 ; python_version >= "3.8" and python_version < "4.0" and sys_platform == "win32"
urllib3==1.26.13 ; python_version >= "3.8" and python_version < "4"
whitenoise==6.3.0 ; python_version >= "3.8" and python_version < "4.0"
